created aggregator for the InstantAction monitoring
